Added optional combo box to select a film to add the actor directly to
Added film title combo box to select a film the actor is in if
applicable. Updated controller to add functionality where if a film is
selected, it also uses a transaction to add a row to the film_actor
junction table. Added new check to ensure the same actor can not be
added more than once.